+++ src/components/domain/Footer/index.ts
@@ -1,1 +1,1 @@
-export * from './Footer';
+export { default as Footer } from './Footer';
