+++ src/components/domain/index.ts
@@ -1,1 +1,2 @@
 export * from './PostItem';
+export * from './Footer';
